+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] <!-- release-date -->
 
+### Changed
+
+- Refuse to clone if repository is not empty.
+- Do not trust the cloned repository.
+
 ## [0.1.2] - 2024-03-29
 
 ### Added