+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] <!-- release-date -->
 
+### Added
+
+- Add unit tests.
+
 ### Changed
 
 - Refuse to clone if repository is not empty.
@@ -39,8 +43,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial implementation in Berlin Rust Hack and Learn 2024-03-28.
 
 <!-- next-url -->
-[Unreleased]: https://github.com/LeoniePhiline/twir-qotw-scraper/compare/v0.1.2...HEAD
 
+[Unreleased]: https://github.com/LeoniePhiline/twir-qotw-scraper/compare/v0.1.2...HEAD
 [0.1.2]: https://github.com/LeoniePhiline/twir-qotw-scraper/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/LeoniePhiline/twir-qotw-scraper/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/LeoniePhiline/twir-qotw-scraper/releases/tag/v0.1.0